Attempt to fix issues
Everything seems to be working expect for the fact that once you create a new event, it won't enter the date data into the database. I think it's an issue with the format of the string given to the backend. There seems to be no easy or convenient method to use to get it into the format that is needed. Next time when I work on this I will have to format the string myself using a function most likely.